requirements.txt
# Penguin Overlord Discord Bot Requirements

# Core Discord.py library
discord.py==2.6.4

# HTTP requests for API calls (XKCD, etc.)
requests==2.32.5

# Environment variable management
python-dotenv==1.2.1

# Secrets Management (choose what you need)
dopplersdk==1.3.0  # For Doppler secrets manager
#boto3==1.28.0  # For AWS Secrets Manager (optional)
#hvac==1.2.0  # For HashiCorp Vault (optional)

# Optional: For better date/time handling
python-dateutil==2.9.0.post0

# Optional: For async HTTP requests (future enhancements)
aiohttp==3.13.2
